chore: rename fluxQueryBuilder component, and fluxQueryBuilder context, into language agnostic
