+++ src/dataExplorer/components/ScriptQueryBuilder.scss
